fix: change several constants to double

+++ include/const.h
@@ -6,12 +6,12 @@
 
 namespace qalsh_chamfer {
 
-const float kEpsilon = 1e-6;
+const double kEpsilon = 1e-6;
 
 const std::string kDefaultParentDirectory = "data";
 const int kDefaultLeftBoundary = -128;  // For dataset generation
 const int kDefaultRightBoundary = 128;  // For dataset generation
-const float kDefaultErrorProbability = 1.0 / std::numbers::e_v<float>;
+const double kDefaultErrorProbability = 1.0 / std::numbers::e_v<double>;
 
 }  // namespace qalsh_chamfer
